trocr_gradio/requirements.txt
gradio
torch
transformers
opencv-python-headless
pillow
numpy